Moved select HOCs UTs under the hoc folder
