+++ packages/react-vapor/src/components/select/hoc/tests/MultiSelectWithFilter.spec.tsx
@@ -8,10 +8,10 @@ import {clearState} from '../../../../utils/ReduxUtils';
 import {TestUtils} from '../../../../utils/tests/TestUtils';
 import {filterThrough} from '../../../filterBox/FilterBoxActions';
 import {ItemBox} from '../../../itemBox/ItemBox';
-import {ISelectWithFilterProps} from '../../hoc/SelectWithFilter';
 import {toggleSelect} from '../../SelectActions';
 import {MultiSelectWithFilter} from '../../SelectComponents';
 import {SelectConnected} from '../../SelectConnected';
+import {ISelectWithFilterProps} from '../SelectWithFilter';
 
 describe('Select', () => {
     describe('MultiSelectWithFilter', () => {
